+++ code/core/src/main/java/org/betonquest/betonquest/item/handler/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Item type handlers and Handler classes.
+ */
+package org.betonquest.betonquest.item.handler;
