+++ code/core/src/main/java/org/betonquest/betonquest/item/typehandler/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Item type handlers.
+ * Item type handler implementations.
  */
 package org.betonquest.betonquest.item.typehandler;
